puhs

+++ ordersts/ordermobile/views.py
@@ -39,7 +39,7 @@ class CashbackMobile(APIView):
                 product_obj = Product.objects.get(id=product["id"])
                 cashback_setting = CashBackSetting.objects.filter(product=product_obj).first()
                 if cashback_setting:
-                    berialadigan_cashback += cashback_setting.cashback_foiz * product["count"] * product_obj.price *doller_value *0.01
+                    berialadigan_cashback += int(cashback_setting.cashback_foiz * product["count"] * product_obj.price *doller_value * 0.01 )
                 
                 else:
                     prod = Product.objects.get(id=product["id"])
